pyproject.toml
[tool.poetry]
name = "anyo-ilora-dllm2rec"
version = "0.1.0"
description = "LLM->Sequential Rec distillation framework (iLoRA + DLLM2Rec)"
authors = ["Gemini"]
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
torch = "==2.3.0"
transformers = "==4.43.0"
hydra-core = "^1.3"
numpy = "^1.26"
pandas = "^2.2"
tensorboard = "^2.18"
pytorch-lightning = "==1.8.6"
bitsandbytes = "==0.37.2"
accelerate = "==0.18.0"
datasets = "==2.10.1"
peft = "==0.3.0"
sentencepiece = "==0.1.97"
scikit-learn = "==1.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8"
black = "^24"
ruff = "^0.7"

[tool.poetry.scripts]
run-teacher = "src.exp.run_teacher:main"
run-student-baseline = "src.exp.run_student_baseline:main"
run-distill = "src.exp.run_distill:main"
run-eval-all = "src.exp.run_eval_all:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
